Proyecto DAW - 04-05-2014
Actualización tras un tiempo sin hacerlo.
Reestructurada casi toda la aplicación usando namespaces.
Cambiada la portada de la página web.
Añadido seguridad a la comunicación AJAX.
Añadida la posibilidad de descargar.

+++ PHP/borraFichero.php
@@ -1,11 +1,20 @@
 <?php session_start();
-    include "funciones.php";
-    $correo =  $_SESSION['nom'];
-    $carpetaActual = $_SESSION['carpetaActual'];
-    $nombreFich = $_POST['nombreFich'];
+	
+	$ajax = $_POST['ajax'];
+    if($ajax == "ajax")
+    {
+	    include "funciones.php";
+	    $correo =  $_SESSION['nom'];
+	    $carpetaActual = $_SESSION['carpetaActual'];
+	    $nombreFich = $_POST['nombreFich'];
 
-    if(unlink('../Usuarios/'.str_replace(".","_",$_SESSION['nom']).$carpetaActual."/".$nombreFich))
-        echo 'SI - ../Usuarios/'.str_replace(".","_",$_SESSION['nom']).$carpetaActual."/".$nombreFich;
-    else echo 'NO - ../Usuarios/'.str_replace(".","_",$_SESSION['nom']).$carpetaActual."/".$nombreFich;
+	    if(unlink('../Usuarios/'.str_replace(".","_",$_SESSION['nom']).$carpetaActual."/".$nombreFich))
+	        echo 'SI - ../Usuarios/'.str_replace(".","_",$_SESSION['nom']).$carpetaActual."/".$nombreFich;
+	    else echo 'NO - ../Usuarios/'.str_replace(".","_",$_SESSION['nom']).$carpetaActual."/".$nombreFich;
+    }
+    else
+    {
+        header( 'Location: Index.php' ) ;
+    }
 
 ?>
